Remove unused CAS3 reports api endpoints
