+++ src/main/resources/static/codegen/built-api-spec.yml
@@ -3903,126 +3903,6 @@ paths:
           $ref: '#/components/responses/403Response'
         500:
           $ref: '#/components/responses/500Response'
-  /reports/bookings:
-    get:
-      tags:
-        - Reports
-      summary: Returns a spreadsheet of all bookings for the specified service and (optional) region.
-      operationId: reportsBookingsGet
-      parameters:
-        - name: X-Service-Name
-          in: header
-          required: true
-          description: Only bookings for this service will be returned
-          schema:
-            $ref: '#/components/schemas/ServiceName'
-        - name: probationRegionId
-          in: query
-          required: false
-          description: If provided, only bookings for this region will be returned
-          schema:
-            type: string
-            format: uuid
-        - name: year
-          in: query
-          required: true
-          description: If provided, only bookings for this year will be returned
-          schema:
-            type: integer
-        - name: month
-          in: query
-          required: true
-          description: If provided, only bookings for this month will be returned - must be provided with year
-          schema:
-            type: integer
-      responses:
-        200:
-          description: Successfully retrieved the report
-          content:
-            'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet':
-              schema:
-                type: string
-                format: binary
-  /reports/bed-usage:
-    get:
-      tags:
-        - Reports
-      summary: Returns a spreadsheet of the bookings, voids and turnarounds all beds for the specified service and (optional) region.
-      operationId: reportsBedUsageGet
-      parameters:
-        - name: X-Service-Name
-          in: header
-          required: true
-          description: Only bookings for this service will be returned
-          schema:
-            $ref: '#/components/schemas/ServiceName'
-        - name: probationRegionId
-          in: query
-          required: false
-          description: If provided, only bookings for this region will be returned
-          schema:
-            type: string
-            format: uuid
-        - name: year
-          in: query
-          required: true
-          description: Only bookings for this year will be returned
-          schema:
-            type: integer
-        - name: month
-          in: query
-          required: true
-          description: Only bookings for this month will be returned - must be provided with year
-          schema:
-            type: integer
-      responses:
-        200:
-          description: Successfully retrieved the report
-          content:
-            'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet':
-              schema:
-                type: string
-                format: binary
-  /reports/bed-utilisation:
-    get:
-      tags:
-        - Reports
-      summary: Returns a spreadsheet showing how many days per month each bedspace spent in which state and the overall occupancy rate.
-      operationId: reportsBedUtilisationGet
-      parameters:
-        - name: X-Service-Name
-          in: header
-          required: true
-          description: Only bookings for this service will be returned
-          schema:
-            $ref: '#/components/schemas/ServiceName'
-        - name: probationRegionId
-          in: query
-          required: false
-          description: If provided, only bookings for this region will be returned
-          schema:
-            type: string
-            format: uuid
-        - name: year
-          in: query
-          required: true
-          description: Only bookings for this year will be returned
-          schema:
-            type: integer
-        - name: month
-          in: query
-          required: true
-          description: Only bookings for this month will be returned - must be provided with year
-          schema:
-            type: integer
-      responses:
-        200:
-          description: Successfully retrieved the report
-          content:
-            'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet':
-              schema:
-                type: string
-                format: binary
 components:
   responses:
     401Response:
